# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BHRM Teams API
  version: 2.0.0
  description: |
    Enterprise Task-Centric Execution Platform — REST API.

    ## Authentication
    All protected routes require `Authorization: Bearer <token>` obtained from
    `POST /auth/login` or `POST /auth/refresh`.

    ## Response envelopes
    | Type | Shape |
    |---|---|
    | Single resource | `{ "success": true, "data": { … } }` |
    | Collection | `{ "success": true, "data": [ … ], "meta": { total, page, limit, totalPages } }` |
    | Error | `{ "success": false, "error": { "code": "…", "message": "…", "statusCode": 4xx } }` |

    ## Role model
    | Role | Description |
    |---|---|
    | `admin` | Company owner — full access to all company resources |
    | `staff` | Invited member — access scoped to assigned tasks |

    ## Endpoint naming conventions
    - Resource-based URLs, nouns only
    - State transitions via `PATCH /{resource}/{id}/status`
    - Bulk assignment via `PATCH /{resource}/{id}/assignees`
    - Avoid custom action verbs in URLs
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    ProjectStatus:
      type: string
      enum: [active, archived]
    SprintStatus:
      type: string
      enum: [draft, active, completed]
    TaskStatus:
      type: string
      enum: [todo, in_progress, review, done]
    TaskPriority:
      type: string
      enum: [low, medium, high, critical]
...
    ActivityAction:
      type: string
      enum: [task_created, task_updated, task_deleted, task_status_changed, task_assigned, task_unassigned, comment_added, comment_edited, comment_deleted, note_created, note_updated, note_deleted, attachment_uploaded, attachment_deleted, step_checked, step_unchecked, sprint_started, sprint_completed, project_archived, staff_invited, staff_suspended, staff_activated]

    # ── Pagination ────────────────────────────────────────────────────────────
    PaginatedMeta:
      type: object
      properties:
        total:      { type: integer, example: 42 }
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        totalPages: { type: integer, example: 3 }

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code:       { type: string, example: NOT_FOUND }
            message:    { type: string, example: Resource not found }
            statusCode: { type: integer, example: 404 }

    # ── User ─────────────────────────────────────────────────────────────────
    UserSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        fullName:  { type: string, example: Jane Smith }
        email:     { type: string, format: email }
        avatarUrl: { type: string, format: uri, nullable: true }
...
    Company:
      type: object
      properties:
        id:                { type: string, format: uuid }
        name:              { type: string, example: Acme Corp }
        slug:              { type: string, example: acme-corp-a3f9b2 }
        logoUrl:           { type: string, format: uri, nullable: true }
        workingHoursStart: { type: string, example: '09:00' }
        workingHoursEnd:   { type: string, example: '18:00' }
        ownerId:           { type: string, format: uuid }
        createdAt:         { type: string, format: date-time }
        updatedAt:         { type: string, format: date-time }
...
    Project:
      type: object
      properties:
        id:          { type: string, format: uuid }
        name:        { type: string, example: Platform V2 }
        description: { type: string, nullable: true }
        status:      { $ref: '#/components/schemas/ProjectStatus' }
        companyId:   { type: string, format: uuid }
        createdBy:   { $ref: '#/components/schemas/UserSummary' }
        sprintCount: { type: integer }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    ProjectDashboard:
      type: object
      properties:
        project:     { $ref: '#/components/schemas/Project' }
        activeSprint:
          nullable: true
          $ref: '#/components/schemas/Sprint'
        tasks:       { $ref: '#/components/schemas/TaskStatusBreakdown' }
        staffCount:  { type: integer }
        recentActivity:
          type: array
          items: { $ref: '#/components/schemas/ActivityItem' }
...
    Sprint:
      type: object
      properties:
        id:        { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        name:      { type: string, example: Sprint 1 }
        goal:      { type: string, nullable: true }
        status:    { $ref: '#/components/schemas/SprintStatus' }
        startDate: { type: string, format: date, nullable: true }
        endDate:   { type: string, format: date, nullable: true }
        taskCount: { type: integer }
        createdAt: { type: string, format: date-time }
...
    StepProgress:
      type: object
      properties:
        total:      { type: integer }
        completed:  { type: integer }
        percentage: { type: integer, minimum: 0, maximum: 100 }
    TaskStatusBreakdown:
      type: object
      properties:
        todo:       { type: integer }
        inProgress: { type: integer }
        review:     { type: integer }
        done:       { type: integer }
        total:      { type: integer }
        overdue:    { type: integer }
    Task:
      type: object
      properties:
        id:                { type: string, format: uuid }
        sprintId:          { type: string, format: uuid }
        parentTaskId:      { type: string, format: uuid, nullable: true }
        name:              { type: string }
        description:       { type: string, nullable: true }
        status:            { $ref: '#/components/schemas/TaskStatus' }
        priority:          { $ref: '#/components/schemas/TaskPriority' }
        startDate:         { type: string, format: date, nullable: true }
        plannedDueDate:    { type: string, format: date, nullable: true }
        actualDueDate:     { type: string, format: date, nullable: true }
        plannedEffortPh:   { type: number }
        estimatedEffortPh: { type: number }
        actualEffortPh:    { type: number }
        slippagePh:
          type: number
          description: 'actualEffortPh − plannedEffortPh; computed on read, never stored'
        isOverdue:         { type: boolean }
        stepProgress:      { $ref: '#/components/schemas/StepProgress' }
        owner:             { $ref: '#/components/schemas/UserSummary' }
        assignees:
          type: array
          items: { $ref: '#/components/schemas/UserSummary' }
        subTaskCount: { type: integer }
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
...
    Note:
      type: object
      properties:
        id:        { type: string, format: uuid }
        taskId:    { type: string, format: uuid }
        title:     { type: string, nullable: true }
        content:   { type: string }
        author:    { $ref: '#/components/schemas/UserSummary' }
        versionCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NoteVersion:
      type: object
      properties:
        id:        { type: string, format: uuid }
        noteId:    { type: string, format: uuid }
        content:   { type: string }
        version:   { type: integer }
        createdBy: { $ref: '#/components/schemas/UserSummary' }
        createdAt: { type: string, format: date-time }
    NoteWithVersions:
      allOf:
        - $ref: '#/components/schemas/Note'
        - type: object
          properties:
            versions:
              type: array
              items: { $ref: '#/components/schemas/NoteVersion' }

    # ── Comment ───────────────────────────────────────────────────────────────
...
    ActivityItem:
      type: object
      properties:
        id:         { type: string, format: uuid }
        action:     { $ref: '#/components/schemas/ActivityAction' }
        entityType:
          type: string
          enum: [task, sprint, project, comment, note, attachment, step, staff, company]
        entityId:   { type: string, format: uuid }
        entityName: { type: string, nullable: true }
        actor:      { $ref: '#/components/schemas/UserSummary', nullable: true }
        metadata:   { type: object, additionalProperties: true, nullable: true }
        createdAt:  { type: string, format: date-time }

    # ── Notification ──────────────────────────────────────────────────────────
...
    AdminDashboard:
      type: object
      properties:
        staff:
          type: object
          properties:
            total:   { type: integer }
            active:  { type: integer }
            pending: { type: integer }
        projects:
          type: object
          properties:
            total:  { type: integer }
            active: { type: integer }
        sprints:
          type: object
          properties:
            active: { type: integer }
        tasks:    { $ref: '#/components/schemas/TaskStatusBreakdown' }
        recentActivity:
          type: array
          items: { $ref: '#/components/schemas/ActivityItem' }
    StaffDashboard:
      type: object
      properties:
        myTasks:      { $ref: '#/components/schemas/TaskStatusBreakdown' }
        upcomingTasks:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        overdueTasks:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        activeSprints:
          type: array
          items: { $ref: '#/components/schemas/Sprint' }
        recentActivity:
          type: array
          items: { $ref: '#/components/schemas/ActivityItem' }
        unreadNotifications: { type: integer }

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
...
    Forbidden:
      description: Insufficient role or resource ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate record or constraint violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    MessageOk:
      description: Operation succeeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean }
              data:
                type: object
                properties:
                  message: { type: string }

  # ── Reusable parameters ────────────────────────────────────────────────────
  parameters:
    companyId:    { name: companyId,    in: path,  required: true, schema: { type: string, format: uuid } }
    projectId:    { name: projectId,    in: path,  required: true, schema: { type: string, format: uuid } }
...
    taskId:       { name: taskId,       in: path,  required: true, schema: { type: string, format: uuid } }
...
    noteId:       { name: noteId,       in: path,  required: true, schema: { type: string, format: uuid } }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
paths:
...
  /companies:
    post:
      tags: [Company]
      summary: Create a company (one per admin account)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, workingHoursStart, workingHoursEnd]
              properties:
                name:              { type: string }
                workingHoursStart: { type: string, pattern: '^\d{2}:\d{2}$', example: '09:00' }
                workingHoursEnd:   { type: string, pattern: '^\d{2}:\d{2}$', example: '18:00' }
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Company' }
        '409': { $ref: '#/components/responses/Conflict' }
  /companies/{companyId}:
    get:
      tags: [Company]
      summary: Get company details
      parameters:
        - $ref: '#/components/parameters/companyId'
      responses:
        '200':
          description: Company
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Company' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Company]
      summary: Update company settings (owner only)
      parameters:
        - $ref: '#/components/parameters/companyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:              { type: string }
                logoUrl:           { type: string, format: uri }
                workingHoursStart: { type: string, pattern: '^\d{2}:\d{2}$' }
                workingHoursEnd:   { type: string, pattern: '^\d{2}:\d{2}$' }
      responses:
        '200':
          description: Updated company
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Company' }
        '400': { description: Invalid working hours range }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Company]
      summary: Soft-delete a company (owner only)
      parameters:
        - $ref: '#/components/parameters/companyId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get project details
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Projects]
      summary: Update project name or description (admin only)
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string }
                description: { type: string }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Project' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /projects/{projectId}/dashboard:
    get:
      tags: [Projects]
      summary: Project dashboard — active sprint, task breakdown, recent activity
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/ProjectDashboard' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tasks/{taskId}/notes:
    get:
      tags: [Notes]
      summary: List notes for a task (newest first)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Paginated notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Note' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Notes]
      summary: Create a note (assignee or admin)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                title:   { type: string, maxLength: 500 }
                content: { type: string, minLength: 1 }
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Note' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /tasks/{taskId}/notes/{noteId}:
    get:
      tags: [Notes]
      summary: Get a note with full version history
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/noteId'
      responses:
        '200':
          description: Note with versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/NoteWithVersions' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Notes]
      summary: Update a note — previous content is auto-snapshotted as a version (admin or author)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/noteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:   { type: string, maxLength: 500 }
                content: { type: string, minLength: 1 }
      responses:
        '200':
          description: Updated note
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Note' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Notes]
      summary: Delete a note (admin or author)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/noteId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════════════════════
  # COMMENTS
  # ════════════════════════════════════════════════════════════════════════════
...
  /dashboard/admin/{companyId}:
    get:
      tags: [Dashboard]
      summary: Admin company dashboard — staff counts, task breakdown, recent activity
      parameters:
        - $ref: '#/components/parameters/companyId'
      responses:
        '200':
          description: Admin dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/AdminDashboard' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /dashboard/me:
    get:
      tags: [Dashboard]
      summary: Staff personal dashboard — task counts, upcoming tasks, active sprints
      responses:
        '200':
          description: Staff dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/StaffDashboard' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /dashboard/projects/{projectId}:
    get:
      tags: [Dashboard]
      summary: Project summary widget (admin only)
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/ProjectDashboard' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════════════════════
  # NOTIFICATIONS
  # ════════════════════════════════════════════════════════════════════════════
...